@@ -1,28 +1,33 @@
 package com.apitest.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
 
 import com.apitest.exception.ParameterException;
 import com.apitest.service.ReturnNumberService;
 
-@RestController
-public class NumberController {
-	    
-	@Autowired
-	private ReturnNumberService reponseService;
-		
-	@GetMapping(value = "/number/{id}", produces = "application/json")
-	public ResponseEntity<String> getnumber(@PathVariable(name = "id", required = false) String id) {
+import lombok.AllArgsConstructor;
 
-		if (Integer.parseInt(id) < 3) {	
-			//throw new ParameterException("Le paramètre 'id' ne doit pas être inferieur à 3");
+@RestController
+@AllArgsConstructor
+public class NumberController {
+
+	private final ReturnNumberService reponseService;
+
+	private final RestTemplate restTemplate = new RestTemplate();
+
+	@GetMapping(value = "/number/{id}", produces = "application/json")
+	public ResponseEntity<String> getnumber(@PathVariable(name = "id", required = false) String id, String token)
+			throws ParameterException {
+
+		if (Integer.parseInt(id) < 3) {
+			throw new ParameterException("Le paramètre 'id' ne doit pas être inferieur à 3");
 		}
-		
+
 		if (id == null) {
 			return new ResponseEntity<>("Aucun 'id' fourni.", HttpStatus.BAD_REQUEST);
 		}
@@ -32,6 +37,11 @@ public class NumberController {
 		}
 
 		return new ResponseEntity<>(reponseService.getNumbers(Integer.parseInt(id)), HttpStatus.OK);
+	}
+
+	@GetMapping("/")
+	public String getGitHub() {
+		return "Welcome, GitHub user";
 	}
 
 	private boolean isInteger(String str) {
