@@ -2,7 +2,6 @@ package com.apitest.exception;
 
 import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -10,11 +9,10 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class CustomErrorController implements ErrorController {
 
-    @RequestMapping("/error")
-    public String handleError() {
-    	log.info("Erreur ");
-        return "error";
-    }
+//    @RequestMapping("/error")
+//    public String handleError() {
+//    	log.info("Erreur ");
+//        return "error";
+//    }
 
-  
 }
